Add Num class test functionality
